requirements.txt
opencv-python
Pillow
python-dotenv
google-generativeai